handle more edge cases during uninstallation

+++ installer/src/main/java/com/teamscale/profiler/installer/steps/InstallSystemdStep.java
@@ -65,6 +65,10 @@ public class InstallSystemdStep implements IStep {
 		}
 
 		Path systemdConfigFile = getSystemdConfigFile();
+		if (!Files.exists(systemdConfigFile)) {
+			return;
+		}
+
 		try {
 			Files.delete(systemdConfigFile);
 		} catch (IOException e) {
